requirements.txt
Django==2.1
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
gunicorn==19.9.0
psycopg2==2.7.5